requirements.txt
Django==4.1
djangorestframework==3.13.1
mysqlclient==2.1.1
djangorestframework-simplejwt==5.2.0
django-dirtyfields==1.8.2
pre-commit==2.20.0
django-cors-headers==3.13.0
python-dotenv==0.19.1
cloudinary==1.28.0
requests~=2.28.1